scaling-agents/DRQN/requirements.txt
pytorch
gymnasium
tensorflow